18 in the bag
Keep on working no matter how many times you fall, keep getting back up.
