+++ build.gradle
@@ -17,7 +17,7 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/TestNG.xml"
+        suites "src/test/resources/Homework18RA.xml"
         systemProperty "browser", System.getProperty("browser")
     }
 }
