+++ src/test/resources/Homework18RA.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="Homework18">
+<suite name="Homework18RA">
     <parameter name="baseURL" value="https://bbb.testpro.io" />
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Homework18">
+            <class name="Homework18RA">
                 <methods>
                     <include name="playSong" />
                 </methods>
