+++ src/test/resources/TestNG.xml
@@ -7,6 +7,7 @@
         <classes>
             <class name="LoginTests"/>
         </classes>
+        </test>
     </test>
 </suite>
