+++ src/test/resources/smokeSuite.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="SmokeSuite">
+<suite name="smokeSuite">
 
     <test name="Smoke Test" preserve-order="false">
         <classes>
             <class name="LoginTests">
                 <methods>
-                    <include name="LoginValidEmailValidPasswordTest" />
+                    <include name="LoginValidEmailPasswordTest" />
                 </methods>
             </class>
         </classes>
     </test>
-
 </suite>
+
